+++ sphinx/source/changelog.rst
@@ -12,6 +12,8 @@ Changelog
 * Added the :ref:`bank` directive (for specifying the RAM bank that is mapped
   to 49152-65535, and for populating a RAM bank from the contents of another
   skool file)
+* Added the ``--banks`` option to :ref:`skool2bin.py` (for processing
+  :ref:`bank` directives and writing RAM banks 0-7 to a 128K file)
 * Added the ``execint`` parameter to the :ref:`AUDIO`, :ref:`SIM` and
   :ref:`TSTATES` macros (to specify whether to simulate the execution of
   interrupt routines)